backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
langgraph==0.2.28
langchain-groq==0.2.0
langchain-core==0.3.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.0